refactor(documentos): collapse 3 triplicate pages into DocumentosModule
Applies the cross-empresa module pattern (like EmpleadosModule in #53) to
the documentos feature. Eliminates ~3,400 LOC of near-duplicated page code
across /administracion/documentos, /dilesa/admin/documentos, and
/rdb/admin/documentos.

Preserves the signed-URL rewriter flow for the private `adjuntos` bucket:
writes store only object paths, and fetchAdjuntosBulk enriches rows with
short-lived signed URLs via lib/adjuntos.ts (getAdjuntoPath +
getAdjuntoSignedUrls). Single-empresa mounts also scope the detail-sheet
update by empresa_id as defense-in-depth (DILESA/RDB previously did this
inline).

New structure under components/documentos/:
 - types.ts, helpers.ts, ui.tsx                  (shared)
 - documento-adjuntos.tsx                        (upload + list)
 - documento-subtipo-fields.tsx                  (escritura/contrato/etc.)
 - documento-form-fields.tsx                     (shared form body)
 - documento-create-sheet.tsx
 - documento-detail-sheet.tsx
 - documentos-table.tsx
 - documentos-module.tsx                         (orchestrator)

Verification: typecheck clean, 0 lint errors (pre-existing warnings only),
format clean, 131/131 tests pass, build compiles all 3 routes.
